pyproject.toml
[project]
name = "careerpilot-ai"
version = "0.1.0"
description = "Agentic AI Career Advisor (LangGraph multi-agent demo)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.14",
    "langgraph>=0.6.6",
    "numpy>=2.0.0",
    "python-dotenv>=1.1.1",
]

